signatory-dalek: remove Ed25519ph (DigestSigner/DigestVerifier)
Removes support for the Ed25519ph algorithm, which accepts a precomputed
SHA-512 digest.

Unfortunately `ed25519-dalek` (and vicariously, `curve25519-dalek`) both
depend on `digest` v0.8, making this feature incompatible with `digest`
and `sha2` upgrades.

Fortunately, to my knowledge nothing actually uses this feature (or for
that matter, Ed25519ph in general) so this isn't a practical problem.