Rewrite as signature KeyRing/KeyStore library
Closes #71, which notes that a lot of the previous functionality of this
library has been extracted into RustCrypto libs or incorporated into the
crates it used to provide wrappers for, including `ed25519-dalek`. The
issue enquires as to what the future of Signatory should be.

This commit completely replaces the existing codebase with the
groundwork for a much higher level signature library based around the
following concepts:

- `KeyRing`: provides a data structure for storing multiple signature
  keys while also spanning multiple signature algorithms
- `KeyStore`: (as a prospective trait) provides storage for keys which
  can be added to a `KeyRing` prior to use

The initial `KeyStore` is a filesystem-backed `FsKeyStore` which manages
PKCS#8-serialized keys. PKCS#8 is a commonly used standard for private
key storage, and includes support for PKCS#5 password-encrypted keys.

This commit adds initial support for ECDSA/secp256k1 as well as an
initial filesystem-backed keystore which can store and load PKCS#8 keys.