Add `Algorithm`, `KeyInfo`, and `KeyName` types
Adds support for querying metadata about a key in the keystore in the
form of a `KeyInfo` struct and an `FsKeyStore::info` method.

Metadata includes the key name, algorithm, and whether the key is
encrypted under a password.

Additionally adds a first-class type for `KeyName` which enforces a
naming invariant, and an `Algorithm` enum with variants for supported
algorithms, selected by which ones are enabled via conditional
compilation.